requirements.txt
playwright
python-telegram-bot==20.3
urllib3==2.0.7
